+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Action to monitor charts
 - Recent Artifacts tab
+- Attach debugger action
 
 ## [2023.1.0] - 2023-01-19